pyproject.toml
[tool.poetry]
name = "managementservice"
version = "1.0.0"
description = "internship project"
authors = ["daniil.prakapchuk@innowise.com"]
readme = "README.md"
packages = []
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.2"
annotated-types = "^0.6.0"
anyio = "^4.3.0"
build = "^1.2.1"
cachecontrol = "^0.14.0"
certifi = "^2024.2.2"
cffi = "^1.16.0"
charset-normalizer = "^3.3.2"
cleo = "^2.1.0"
click = "^8.1.7"
crashtest = "^0.4.1"
cryptography = "^42.0.5"
distlib = "^0.3.8"
dulwich = "^0.21.7"
fastjsonschema = "^2.19.1"
filelock = "^3.13.4"
h11 = "^0.14.0"
httptools = "^0.6.1"
idna = "^3.7"
installer = "^0.7.0"
jaraco-classes = "^3.4.0"
jeepney = "^0.8.0"
keyring = "^25.1.0"
more-itertools = "^10.2.0"
msgpack = "^1.0.8"
packaging = "^24.0"
pexpect = "^4.9.0"
pkginfo = "^1.10.0"
platformdirs = "^4.2.0"
poetry-core = "^1.9.0"
ptyprocess = "^0.7.0"
pycparser = "^2.22"
pydantic = {extras = ["email"], version = "^2.7.0"}
pydantic-core = "^2.18.1"
pyproject-hooks = "^1.0.0"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"
rapidfuzz = "^3.8.1"
requests = "^2.31.0"
requests-toolbelt = "^1.0.0"
secretstorage = "^3.3.3"
shellingham = "^1.5.4"
sniffio = "^1.3.1"
starlette = "^0.37.2"
tomlkit = "^0.12.4"
trove-classifiers = "^2024.4.10"
typing-extensions = "^4.11.0"
urllib3 = "^2.2.1"
uvicorn = "^0.29.0"
uvloop = "^0.19.0"
virtualenv = "^20.25.3"
watchfiles = "^0.21.0"
websockets = "^12.0"
pydantic-extra-types = "^2.6.0"
sqlalchemy = "^2.0.29"
aioredis = "^2.0.1"
phonenumbers = "^8.13.35"
asyncpg = "^0.29.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pika = "^1.3.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pycryptodome = "^3.20.0"
python-multipart = "^0.0.9"
pydantic-settings = "^2.2.1"
python-magic = "^0.4.27"
aioboto3 = "^12.4.0"
libmagic = "^1.0"
ruff = "^0.4.1"
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"
faker = "^24.11.0"
httpx = "^0.27.0"
testcontainers = "^4.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.format]
quote-style = "single"